requirements.txt
pymupdf
langchain-text-splitters
python-dotenv
google-generativeai
chromadb
streamlit>=1.30.0
